Some improvements in the annealing
Now the cost function is not based anymore on minimizing the dissimilarity between the orders of distances ; this was source of imprecisions, since order relationships are not sensitive to magnitude of the distances. On the opposite now, both the distance among signals and distances among position in the memory are rescaled in order to remove "unities of measure".
Doing this we can compare directly the two matrices in order to search for the configuration that minimizes the difference